Include all events in the reference guide
Responds to zmb3 feedback.

When we add a new event, we need to edit the formatters in
`makeEvent.ts` so the event appears in the Web UI. For any event
formatter with no test fixture, include the event, type, code, and
description in the reference guide using data from the formatter, but
don't include an example. Also remove any test fixtures from the guide
that don't correspond to a formatter.

Also log events with no examples when running the generator so we can
add test fixtures later on.