Add AWS CredentialProcess credentials field to RouteToApp
RouteToApp will now contain the credentials to access AWS APIs.
This is only valid for AWS Access paths that use an AWS IAM Roles
Anywhere Integration.

These credentials are stored in the user certificate.

`tsh` will be used as the `credential_process` for a profile, and will
return the credentials as they exist in the app's certificate.